// webapp/src/modules/vendor/decentraland/routing/types.ts
export enum Section {
  ALL = 'all',
  POPULAR = 'popular',
  NEWEST = 'newest',
  LATEST_SOLD = 'latest_sold',
  GENERATIONS = 'generations',
  PRICE = 'price',
  GENERATIONS_0 = 'gen_0',
  GENERATIONS_1 = 'gen_1',
  GENERATIONS_2 = 'gen_2',
  GENERATIONS_3 = 'gen_3',
  GENERATIONS_4 = 'gen_4',
  GENERATIONS_5 = 'gen_5',
  GENERATIONS_6 = 'gen_6',
  GENERATIONS_7 = 'gen_7',
  GENERATIONS_8 = 'gen_8',
  GENERATIONS_9 = 'gen_9',

  LAND = 'land',
  PARCELS = 'parcels',
  ESTATES = 'estates',
  WEARABLES = 'wearables',
  KRYPTOMONS = 'kryptomons',
  WEARABLES_HEAD = 'wearables_head',
  WEARABLES_EYEBROWS = 'wearables_eyebrows',
  WEARABLES_EYES = 'wearables_eyes',
  WEARABLES_FACIAL_HAIR = 'wearables_facial_hair',
  WEARABLES_HAIR = 'wearables_hair',
  WEARABLES_MOUTH = 'wearables_mouth',

  WEARABLES_UPPER_BODY = 'wearables_upper_body',
  WEARABLES_LOWER_BODY = 'wearables_lower_body',
  WEARABLES_FEET = 'wearables_feet',

  WEARABLES_ACCESORIES = 'wearables_accesories',
  WEARABLES_EARRING = 'wearables_earring',
  WEARABLES_EYEWEAR = 'wearables_eyewear',
  WEARABLES_HAT = 'wearables_hat',
  WEARABLES_HELMET = 'wearables_helmet',
  WEARABLES_MASK = 'wearables_mask',
  WEARABLES_TIARA = 'wearables_tiara',
  WEARABLES_TOP_HEAD = 'wearables_top_head',

  ENS = 'ens',

  ELEMENT_TYPE = 'element_type',
  SPECIALTY = 'specialty',
  SUPER = 'super',
  AFFECTION = 'affection',
  CRAZINESS = 'craziness',
  INSTINCT = 'instinct',
  HUNGER = 'hunger',
  LAZINESS = 'laziness',
  SMART = 'smart',
  ELEMENT_STARTING_TALENT = 'element_starting_talent',
  CONSTITUTION = 'constitution',
  BRAVENESS = 'braveness',
  SEX = 'sex',
  SKIN_TYPE = 'skin_type',
  BODY_SIZE = 'body_size',
  EGO = 'ego',
  HEALTH_POINTS = 'health_points',
  SPEED = 'speed',
  KRYPTOMON_STATUS = 'kryptomon_status',
  COLOR = 'color',
  ATTACK = 'attack',
  DEFENCE = 'defence',
  GENERAL_TALENT = 'general_talent',
  GROWTH_TALENT_FACTOR = 'growth_talent_factor',
  ELEMENT_PERCENTAGE = 'element_percentage',
  SPECIAL = 'special',

  WATER = 'water',
  WATER_TALENT = 'water_talent',
  FIRE = 'fire',
  FIRE_TALENT = 'fire_talent',
  GROUND = 'ground',
  GROUND_TALENT = 'ground_talent',
  ICE = 'ice',
  ICE_TALENT = 'ice_talent',
  GRASS = 'grass',
  GRASS_TALENT = 'grass_talent',
  ELECTRO = 'electro',
  ELECTRO_TALENT = 'electro_talent',
  GHOST = 'ghost',
  GHOST_TALENT = 'ghost_talent',
  AIR = 'air',
  AIR_TALENT = 'air_talent',
  UNFREEZABLE = 'unfreezable'
}